move scrubber to own file
